cleaned up uswds table
